fdfdf
dfsdfsdf

+++ productos.php
@@ -80,10 +80,21 @@ try{
             print json_encode($jTableResult);
         }
 	}else if($_GET["accion"] == "eliminar"){
+        $result = mysql_query("SELECT count(*) as cont FROM gestion where id_empresa=".$_POST["id"].";");
+        $row = mysql_fetch_array($result);
+        if($row['cont'] == 0){
 		$result = mysql_query("DELETE FROM empresa WHERE id= " . $_POST["id"] . ";");
 		$jTableResult = array();
 		$jTableResult['Result'] = "OK";
 		print json_encode($jTableResult);
+        }else{
+            $row = mysql_fetch_array($result);
+            $jTableResult = array();
+            $jTableResult['Result'] = "ERROR";
+            $jTableResult['Message'] = "No puede eliminar teniendo gestiones en la empresa";
+            print json_encode($jTableResult);
+
+        }
 	}
 	mysql_close($con);
